requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
httpx==0.25.0
python-dotenv==1.0.0
requests==2.31.0
chromadb==0.4.15
langchain==0.0.350requests
